darc-el-backend/pyproject.toml
[build-system]
requires = ["hatch>=1.16", "hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "darc-el-backend"
version = "0.1.0"
description = "DARC-EL backend"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Roland Ramp" },
]
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.136.0",
  "httpx>=0.28.1",
  "neo4j>=5.28.1",
  "openai>=2.32.0",
  "pydantic>=2.13.2",
  "pyyaml>=6.0.3",
  "pypdf>=6.10.2",
  "pyzotero>=1.11.0",
  "python-dotenv>=1.2.2",
  "python-multipart>=0.0.26",
  "python-docx>=1.2.0",
  "uvicorn>=0.44.0",
  "openrouter>=0.9.1"
]

[project.scripts]
zotero-download = "main:main"

[project.optional-dependencies]
dev = [
  "ruff>=0.15.11",
]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src", "tests"]
exclude = [
  "src/darc_el.egg-info",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
include = ["src/main.py"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
include = ["src/main.py", "README.md", "pyproject.toml"]
